speak_futbal/backend/events/serializers.py
from rest_framework import serializers
from .models import Event
from users.serializers import UserSerializer

class EventSerializer(serializers.ModelSerializer):
    created_by = UserSerializer(read_only=True)
    participants = UserSerializer(many=True, read_only=True)
    participant_count = serializers.SerializerMethodField()

    class Meta:
        model = Event
        fields = '__all__'
        read_only_fields = ('created_by', 'created_at', 'updated_at')

    def get_participant_count(self, obj):
        return obj.participants.count()

    def create(self, validated_data):
        validated_data['created_by'] = self.context['request'].user
        return super().create(validated_data)

class EventListSerializer(serializers.ModelSerializer):
    created_by = UserSerializer(read_only=True)
    participant_count = serializers.SerializerMethodField()

    class Meta:
        model = Event
        fields = ('id', 'title', 'event_type', 'location', 'start_date', 'end_date', 
                 'created_by', 'participant_count', 'max_participants', 'is_active')

    def get_participant_count(self, obj):
        return obj.participants.count() 